requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
backports.ssl-match-hostname==3.5.0.1
botocore==1.5.47
cement==2.8.2
cffi==1.10.0
colorama==0.3.7
cryptography==1.8.1
dj-database-url==0.4.1
Django==1.8
django-appconf==1.0.2
django-bootstrap-themes==3.1.2
django-bootstrap3==5.4.0
django-compressor==1.5
django-email-bandit==1.2
django-otp==0.3.13
-e git://github.com/bitmazk/django-registration-email@f2e0b08cf4a26dba48be8f87f1e5ce5df916c80c#egg=django_registration_email
django-registration-redux==1.1
django-staticfiles==1.2.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
enum34==1.1.6
gunicorn==19.6.0
idna==2.5
instamojo-wrapper==1.1
ipaddress==1.0.18
jmespath==0.9.2
oauthlib==2.0.2
packaging==16.8
pathspec==0.5.0
paypalrestsdk==1.10.0
pbr==3.0.0
Pillow==2.9.0
psycopg2==2.6.2
pycparser==2.17
PyJWT==1.5.0
pyOpenSSL==17.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-openid==2.2.5
python-social-auth==0.2.7
pytz==2017.2
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.8.0
semantic-version==2.5.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.3.0
sqlparse==0.1.19
stevedore==1.21.0
tabulate==0.7.5
termcolor==1.1.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
virtualenvwrapper-win==1.2.1
websocket-client==0.40.0
whitenoise==2.0.6
